Get rid of sync peewee ORM. We need to use async libs like asyncpgsa.
