+++ alembic.ini
@@ -2,7 +2,7 @@
 
 [alembic]
 # path to migration scripts
-script_location = migration
+script_location = src
 
 # template used to generate migration files
 # file_template = %%(rev)s_%%(slug)s
