--- migration/requirements.txt
@@ -1,3 +1,0 @@
-alembic==0.9.9
-sqlacodegen==2.0.0
-SQLAlchemy==1.2.7
