+++ server/requirements.txt
@@ -1,4 +1,6 @@
 aiohttp==3.1.2
-peewee==3.2.2
-psycopg2==2.7.4
+alembic==0.9.9
+asyncpgsa
 marshmallow==3.0.0b11
+sqlacodegen==2.0.0
+SQLAlchemy==1.2.7
