+++ server/src/env.py
@@ -22,7 +22,7 @@ fileConfig(config.config_file_name)
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
 
-from migration.models import Base
+from server.src.models import Base
 target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
